+++ src/configs/eslint/expo.dependencies.ts
@@ -5,6 +5,6 @@ import { dependencies as reactDependencies } from './react.dependencies.js';
  */
 export const dependencies = {
     ...reactDependencies,
-    '@eslint/compat': '^1',
     'eslint-plugin-react-native': '^5',
+    '@react-native/eslint-config': '^0.85',
 } as const;